+++ tests/env_test.py
@@ -1,11 +1,18 @@
-from .context import preflight as pre
+from .context import preflightpy as pre
+
 
 class TestEnvUpdate:
 
     def test_pressure(self):
-        z = lambda h: 6378137*h/(6378137-h)
-        parameters = pre.Parameters("tests/input/case.json")
-        environment = pre.Environment(parameters.env_variables)
+        environment = pre.Environment([
+            113,
+            0.01,
+            9.80665,
+            0.02896968,
+            8.314462618,
+            1.4,
+            101325
+            ])
         environment.get_status(5000)
         assert round(environment.P, -1) == 54040
         environment.get_status(11019)
@@ -14,10 +21,39 @@ class TestEnvUpdate:
         assert round(environment.P, -1) == 5530
         environment.get_status(70000)
         assert round(environment.P, 1) == 5.2
+        environment.get_status(87000)
+        assert round(environment.P, 2) == 0.31
+        environment.get_status(95000)
+        assert round(environment.P, 3) == 0.076
+        environment.get_status(105000)
+        assert round(environment.P, 3) == 0.014
+        environment.get_status(115000)
+        assert round(environment.P, 4) == 0.0040
+        environment.get_status(140000)
+        assert round(environment.P, 5) == 0.00072
+        environment.get_status(170000)
+        assert round(environment.P, 5) == 0.00021
+        environment.get_status(250000)
+        assert round(environment.P, 6) == 0.000025
+        environment.get_status(400000)
+        assert round(environment.P, 8) == 1.45e-6
+        environment.get_status(625000)
+        assert round(environment.P, 10) == 6.26e-8
+        environment.get_status(800000)
+        assert round(environment.P, 10) == 1.70e-8
+        environment.get_status(1000000)
+        assert round(environment.P, 11) == 7.51e-9
 
     def test_temperature(self):
-        parameters = pre.Parameters("tests/input/case.params")
-        environment = pre.Environment(parameters.env_variables)
+        environment = pre.Environment([
+            113,
+            0.01,
+            9.80665,
+            0.02896968,
+            8.314462618,
+            1.4,
+            101325
+            ])
         environment.get_status(5000)
         assert round(environment.T, 2) == 255.68
         environment.get_status(20063)
@@ -27,11 +63,36 @@ class TestEnvUpdate:
         environment.get_status(47350)
         assert round(environment.T, 2) == 270.65
         environment.get_status(71802)
-        assert round(environment.T, 2) == 214.64
+        assert round(environment.T, 2) == 214.65
+        environment.get_status(87000)
+        assert round(environment.T, 2) == 186.67
+        environment.get_status(100000)
+        assert round(environment.T, 2) == 195.08
+        environment.get_status(115000)
+        assert round(environment.T, 2) == 300.00
+        environment.get_status(125000)
+        assert round(environment.T, 2) == 417.23
+        environment.get_status(160000)
+        assert round(environment.T, 2) == 696.29
+        environment.get_status(330000)
+        assert round(environment.T, 2) == 985.88
+        environment.get_status(500000)
+        assert round(environment.T, 2) == 999.24
+        environment.get_status(750000)
+        assert round(environment.T, 2) == 999.99
+        environment.get_status(1000000)
+        assert round(environment.T, 2) == 1000.00
 
     def test_density(self):
-        parameters = pre.Parameters("tests/input/case.json")
-        environment = pre.Environment(parameters.env_variables)
+        environment = pre.Environment([
+            113,
+            0.01,
+            9.80665,
+            0.02896968,
+            8.314462618,
+            1.4,
+            101325
+            ])
         environment.get_status(5000)
         assert round(environment.Rho, 4) == 0.7365
         environment.get_status(11019)
@@ -41,10 +102,16 @@ class TestEnvUpdate:
         environment.get_status(40000.0)
         assert round(environment.Rho, 4) == 0.0040
 
-
     def test_speed_of_sound(self):
-        parameters = pre.Parameters("tests/input/case.params")
-        environment = pre.Environment(parameters.env_variables)
+        environment = pre.Environment([
+            113,
+            0.01,
+            9.80665,
+            0.02896968,
+            8.314462618,
+            1.4,
+            101325
+            ])
         environment.get_status(1000)
         assert round(environment.c, 1) == 336.4
         environment.get_status(10000)