+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to semantic versioning: [SemVer](https://semver.org/).
 
-## [Unreleased]
+## [0.6.0] - 2019-11-2
+### Added
+- Unit tests for heterosphere model
+- Unit tests for `params.py`
+
+### Changed
+- `env.py` for 86km - 1000km support
+
+## [0.5.0] - 2019-11-1
 ### Added
 - `env.py` for the `Environment` class
 - `case.json` for input